infrastructure/files/cloud_functions/vertex_pipeline_alerting/requirements.txt
google-cloud-aiplatform==1.27.0